bioterm/server/backend/requirements.txt
aiofiles==0.7.0
alabaster==0.7.13
alembic==1.12.0
aniso8601==7.0.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
Authlib==1.2.1
Babel==2.12.1
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
distlib==0.3.7
dnspython==2.3.0
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.0.0.post2
esbonio==0.16.1
exceptiongroup==1.1.2
fastapi==0.104.1
fastapi-third-party-auth==0.1.2
filelock==3.12.2
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lsprotocol==2023.0.0b1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.2
packaging==23.1
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.2.0
promise==2.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pygls==1.0.2
Pygments==2.16.1
pyproject-api==1.5.4
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pythondialog==3.5.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
Rx==1.6.3
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.49
starlette==0.27.0
tomli==2.0.1
tox==4.10.0
typeguard==3.0.2
typing_extensions==4.9.0
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.24.0.post1
uvloop==0.17.0
virtualenv==20.24.3
watchfiles==0.21.0
watchgod==0.8.2
websockets==11.0.3
zipp==3.17.0
